Merge branch 'master' into 1579

# Conflicts:
#	Application/Ringtoets/src/Application.Ringtoets.Migration.Core/EmbeddedResources/DatabaseStructure18.1.sql
#	Application/Ringtoets/src/Application.Ringtoets.Storage/DbContext/RingtoetsEntities.edmx.diagram
#	Application/Ringtoets/src/Application.Ringtoets.Storage/Ringtoets.rtd
#	build/DatabaseStructure.sql
#	design/DatabaseDesign.eap

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/DbContext/StabilityPointStructuresFailureMechanismMetaEntity.cs
@@ -34,7 +34,7 @@ namespace Application.Ringtoets.Storage.DbContext
     {
         public long StabilityPointStructuresFailureMechanismMetaEntityId { get; set; }
         public long FailureMechanismEntityId { get; set; }
-        public int N { get; set; }
+        public double N { get; set; }
         public string ForeshoreProfileCollectionSourcePath { get; set; }
         public string StabilityPointStructureCollectionSourcePath { get; set; }
